+++ glship.cpp
@@ -129,7 +129,7 @@ void GLShip::collide(GLShip* first, GLShip* second) {
   Ship::collide(first->ship, second->ship);
 }
 
-void GLShip::step(float delta, const Grid &grid) {
+void GLShip::step(int delta, const Grid &grid) {
   ship->step(delta, grid);
 
   float camera_rotation_delta = ship->heading() - camera_rotation;
